Implementing full object copy for tree, starting unit test

+++ app/controllers/tree/JPATreeStorage.java
@@ -1,13 +1,18 @@
 package controllers.tree;
 
+import java.math.BigInteger;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import javax.persistence.Query;
+
 import models.tree.GenericTreeNode;
 import models.tree.Node;
+import models.tree.jpa.AbstractNode;
 import models.tree.jpa.TreeNode;
 import play.db.jpa.JPA;
+import play.db.jpa.JPABase;
 import play.db.jpa.Model;
-
-import javax.persistence.Query;
-import java.util.List;
 
 /**
  * @author Manuel Bernhardt <bernhardt.manuel@gmail.com>
@@ -44,50 +49,32 @@ public class JPATreeStorage extends TreeStorage {
 
     @Override
     public GenericTreeNode getTreeNode(Long id) {
-        return (GenericTreeNode) TreeNode.findById(id);
+        JPABase node = TreeNode.findById(id);
+        return (GenericTreeNode) node;
 
     }
 
     @Override
     public void remove(Long id, boolean removeObject) {
-
-        // making this behaviour configurable is some manual labor
-        if (removeObject) {
-            Long oid = getAbstractNodeId(id);
-            removeReferences(id);
-            updateQuery("delete from AbstractNode a where a.id = ?", oid);
-            deleteNode(id, removeObject);
-        } else {
-            updateQuery("update AbstractNode a set a.treeNode = null where a.treeNode.id = ?", id);
-            removeReferences(id);
-            deleteNode(id, removeObject);
-        }
-    }
-
-    private void deleteNode(Long id, boolean removeObject) {
-        // kill the kids
         TreeNode parent = TreeNode.findById(id);
-        Query q = JPA.em().createQuery("select n.id from TreeNode n where n.path like :path and n.id <> :id order by n.path desc");
-        q.setParameter("id", id);
+
         String pathLike = parent.getPath() + "%";
-        q.setParameter("path", pathLike);
-        List<Long> kids = q.getResultList();
-        updateQuery("update TreeNode n set n.abstractNode = null where n.path like ? and n.id <> ?", pathLike, id);
-        if (removeObject) {
-            namedUpdateQuery("delete from AbstractNode a where a.treeNode.id in (:kids)", "kids", kids);
-        } else {
-            namedUpdateQuery("update AbstractNode a set a.treeNode = null where a.treeNode.id in (:kids)", "kids", kids);
+        List<Long> kids = queryList("select n.id from TreeNode n where n.path like ? and n.level > ? and n.threadRoot.id = ? order by n.path desc", Long.class, pathLike, parent.getLevel(), parent.getThreadRoot().getId());
+        if (!kids.isEmpty()) {
+            List<Long> nodes = queryList("select n.abstractNode.id from TreeNode n where n.path like ? and n.level > ? and n.threadRoot.id = ? order by n.path desc", Long.class, pathLike, parent.getLevel(), parent.getThreadRoot().getId());
+            if (removeObject) {
+                namedUpdateQuery("update TreeNode n set n.abstractNode = null where n.id in (:kids)", "kids", kids);
+                namedUpdateQuery("delete from AbstractNode a where a.id in (:nodes)", "nodes", nodes);
+            }
+            namedUpdateQuery("delete from TreeNode n where n.id in (:kids)", "kids", kids);
         }
-        namedUpdateQuery("delete from TreeNode n where n.id in (:kids)", "kids", kids);
-        updateQuery("delete from TreeNode n where n.id = ?", id);
-    }
 
-    private void removeReferences(Long id) {
         updateQuery("update TreeNode n set n.abstractNode = null, n.threadRoot = null where n.id = ?", id);
-    }
 
-    private Long getAbstractNodeId(Long id) {
-        return query("select a.id from TreeNode n join n.abstractNode a where n.id = :id", id, Long.class);
+        if (removeObject) {
+            updateQuery("delete from AbstractNode a where a.id = ?", parent.abstractNode.getId());
+        }
+        updateQuery("delete from TreeNode n where n.id = ?", id);
     }
 
     @Override
@@ -107,7 +94,7 @@ public class JPATreeStorage extends TreeStorage {
     @Override
     public void rename(Long id, String name) {
         updateQuery("update TreeNode n set n.name = ? where n.id = ?", name, id);
-        updateQuery("update AbstractNode n set n.name = ? where n.treeNode.id = ?", name, id);
+        updateQuery("update AbstractNode a set a.name = ? where a.id = (select n.abstractNode.id from TreeNode n where n.id = ?)", name, id);
     }
 
     @Override
@@ -119,7 +106,7 @@ public class JPATreeStorage extends TreeStorage {
         String newPath = parent.getPath();
         Integer delta = parent.getLevel() - node.getLevel() + 1;
 
-        if(node.getThreadRoot().getId() == node.getId()) {
+        if (node.getThreadRoot().getId() == node.getId()) {
             updateQuery("update TreeNode set path = concat(?, path), level = level + ? where threadRoot = ?", newPath + "____", delta, parent.getThreadRoot());
         } else {
             String oldPath = node.getPath();
@@ -130,9 +117,8 @@ public class JPATreeStorage extends TreeStorage {
     }
 
 
-
     @Override
-    public void copy(Long id, Long target, boolean copyObject) {
+    public void copy(Long id, Long target, boolean copyObject, NodeType[] types) {
         TreeNode node = TreeNode.findById(id);
         TreeNode parent = TreeNode.findById(target);
         Integer delta = parent.getLevel() - node.getLevel() + 1;
@@ -141,8 +127,7 @@ public class JPATreeStorage extends TreeStorage {
         Integer oldPathLength = node.getParent().getPath().length();
         String pathLike = oldPath + "%";
 
-        if(!copyObject) {
-            // TODO file hibernate bug
+        // see http://opensource.atlassian.com/projects/hibernate/browse/HHH-2692
 //            Query query = JPA.em().createQuery("insert into TreeNode (name, typeName, opened, level, path, threadRoot, abstractNode) " +
 //                    "select c.name, c.typeName, c.opened, c.level + :delta, concat(:newPath, substring(path, :oldPathLength, length(path))), (select :newThreadRoot from AbstractNode), c.abstractNode " +
 //                    "from TreeNode c " +
@@ -154,23 +139,60 @@ public class JPATreeStorage extends TreeStorage {
 //            query.setParameter("oldThreadRoot", node.getThreadRoot());
 //            query.setParameter("pathLike", pathLike);
 
-            Query query = JPA.em().createNativeQuery("insert into TreeNode (name, typeName, opened, level, path, threadRoot_id, abstractNode_id) " +
-                    "select c.name, c.typeName, c.opened, c.level + ?, concat(?, substring(path, ?, length(path))), ?, c.abstractNode_id " +
-                    "from TreeNode c " +
-                    "where c.threadRoot_id = ? and c.path like ?");
-            query.setParameter(1, delta);
-            query.setParameter(2, newPath);
-            query.setParameter(3, oldPathLength + 1);
-            query.setParameter(4, parent.getThreadRoot().getId());
-            query.setParameter(5, node.getThreadRoot().getId());
-            query.setParameter(6, pathLike);
+        Query query = JPA.em().createNativeQuery("insert into TreeNode (name, typeName, opened, level, path, threadRoot_id, abstractNode_id) " +
+                "select c.name, c.typeName, c.opened, c.level + ?, concat(?, substring(path, ?, length(path))), ?, c.abstractNode_id " +
+                "from TreeNode c " +
+                "where c.threadRoot_id = ? and c.path like ?");
+        query.setParameter(1, delta);
+        query.setParameter(2, newPath);
+        query.setParameter(3, oldPathLength + 1);
+        query.setParameter(4, parent.getThreadRoot().getId());
+        query.setParameter(5, node.getThreadRoot().getId());
+        query.setParameter(6, pathLike);
 
-            query.executeUpdate();
-        } else {
-            // TODO
+        query.executeUpdate();
 
+        if (copyObject) {
+            // fetch all (type, (AbstractNode, TreeNode)) where TreeNode-s are the freshly copied nodes (they still point to the original AbstractNode)
+            Map<String, Map<Long, Long>> nodeIdsByType = new HashMap<String, Map<Long, Long>>();
+            Query copied = JPA.em().createNativeQuery("select n.typeName, n.abstractNode_id, n.id from TreeNode n join TreeNode o where n.abstractNode_id = o.abstractNode_id and n.id <> o.id and n.path not like ?");
+            copied.setParameter(1, pathLike);
+            List<Object[]> ids = (List<Object[]>) copied.getResultList();
+            for (Object[] p : ids) {
+                String type = (String) p[0];
+                Map<Long, Long> i = nodeIdsByType.get(type);
+                if (i == null) {
+                    i = new HashMap<Long, Long>();
+                    nodeIdsByType.put(type, i);
+                }
+                i.put(((BigInteger) p[1]).longValue(), ((BigInteger) p[2]).longValue());
+            }
+
+            // with the inheritance strategy in use we have no choice but to copy the concrete nodes by hand using reflection
+            for (NodeType type : types) {
+                Map<Long, Long> copyIds = new HashMap<Long, Long>();
+                Map<Long, Long> i = nodeIdsByType.get(type.getName());
+                if(i != null) {
+                    try {
+                        Query q = JPA.em().createQuery("from " + type.getNodeClass().getSimpleName() + " n where n.id in (:ids)");
+                        q.setParameter("ids", i.keySet());
+                        List<AbstractNode> original = q.getResultList();
+                        for (AbstractNode n : original) {
+                            AbstractNode copy = (AbstractNode) n.clone();
+                            copy.id = null;
+                            copy.create();
+                            copyIds.put(i.get(n.getId()), copy.getId());
+                        }
+                    } catch (Exception e) {
+                        e.printStackTrace();
+                    }
+                    // TODO see if this can be optimized
+                    for (Long tn : copyIds.values()) {
+                        updateQuery("update TreeNode set abstractNode = (select a.id from AbstractNode a where a.id = ?) where id = ?", copyIds.get(tn), tn);
+                    }
+                }
+            }
         }
-
     }
 
     private <T> T query(String query, Object id, Class<T> type) {
@@ -217,5 +239,29 @@ public class JPATreeStorage extends TreeStorage {
         Query q = JPA.em().createQuery(query);
         q.setParameter(argName, arg);
         q.executeUpdate();
+    }
+
+    private static class CopyId {
+        private final Long id;
+        private final Long abstractNodeId;
+        private final String type;
+
+        private CopyId(Long id, Long abstractNodeId, String type) {
+            this.id = id;
+            this.abstractNodeId = abstractNodeId;
+            this.type = type;
+        }
+
+        public Long getId() {
+            return id;
+        }
+
+        public Long getAbstractNodeId() {
+            return abstractNodeId;
+        }
+
+        public String getType() {
+            return type;
+        }
     }
 }
